feat: [BSZ-MODEL-01-model-service-surfaces-001] - [Preserve leased inference lifecycle behavior in focused tests]
